[FIX] Refatoração do modulo 'project_task_calendar'
