+++ project_task_calendar/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 
 from . import models
-from . import wizard
+# from . import wizard
