+++ project_task_calendar/__manifest__.py
@@ -21,9 +21,9 @@
     'data': [
         # 'security/ir.model.access.csv',
         # 'data/to_invoice_data.xml',
-        'views/view_calendar.xml',
-        'views/view_project_project.xml',
-        'views/view_project_task.xml',
+        'views/calendar_event.xml',
+        'views/project_project.xml',
+        'views/project_task.xml',
         # 'views/to_invoice_view.xml',
         # 'views/account_analytic_line_view.xml',
         # 'wizard/wizard_calendar_event_done.xml',
