+++ project_task_calendar/models/__init__.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
-from . import calendar
-from . import project
-from . import account_analytic_line
+from . import calendar_event
+from . import project_project
+from . import project_task
+# from . import account_analytic_line
 # from . import to_invoice
