+++ project_task_calendar/wizard/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
 
-from . import wizard_calendar_event_done
+# from . import wizard_calendar_event_done
